feat: refactoring dashboard before working on ux details
