--- harp_apps/dashboard/frontend/src/Pages/Overview/Components/Charts/index.ts
@@ -1,1 +1,0 @@
-export { TransactionsChart } from "./TransactionsChart"
