--- harp_apps/dashboard/frontend/src/Pages/Overview/Components/index.ts
@@ -1,1 +1,0 @@
-export { BaseTransactionsOverview } from "./BaseTransactionsOverview"
