--- harp_apps/dashboard/frontend/src/Pages/Overview/Containers/index.ts
@@ -1,1 +1,0 @@
-export { TransactionsOverview } from "./TransactionsOverview"
