+++ harp_apps/dashboard/frontend/src/Pages/Overview/Sections/Components/Charts/index.ts
@@ -1,0 +1,1 @@
+export { TransactionsChart } from "./TransactionsChart.tsx"
