+++ harp_apps/dashboard/frontend/src/Pages/Overview/Sections/Components/TransactionsHistory.test.tsx
@@ -2,14 +2,14 @@ import { ErrorBoundary } from "react-error-boundary"
 import { expect, it } from "vitest"
 
 import { Error } from "Components/Page"
-import { renderWithClient } from "tests/utils"
+import { renderWithClient } from "tests/utils.tsx"
 
-import { TransactionsOverview } from "./TransactionsOverview"
+import { TransactionsHistory } from "./TransactionsHistory.tsx"
 
 it("renders the title and data when the query is successful", async () => {
   const result = renderWithClient(
     <ErrorBoundary FallbackComponent={Error}>
-      <TransactionsOverview endpoint="test" title="Test Title" className="test-class" timeRange="month" />
+      <TransactionsHistory endpoint="test" title="Test Title" className="test-class" timeRange="month" />
     </ErrorBoundary>,
   )
