+++ harp_apps/dashboard/frontend/src/Pages/Overview/Sections/Components/TransactionsHistoryOnQuerySuccess.test.tsx
@@ -1,7 +1,10 @@
 import { render } from "@testing-library/react"
 import { expect, it } from "vitest"
 
-import { BaseTransactionsOverview, TransactionOverviewChartProps } from "./BaseTransactionsOverview"
+import {
+  TransactionOverviewChartProps,
+  TransactionsHistoryOnQuerySuccess,
+} from "./TransactionsHistoryOnQuerySuccess.tsx"
 
 it("renders without crashing", () => {
   const data: TransactionOverviewChartProps["data"] = {
@@ -13,6 +16,8 @@ it("renders without crashing", () => {
     count: 100,
   }
 
-  const { container } = render(<BaseTransactionsOverview data={data} title="Test Title" className="test-class" />)
+  const { container } = render(
+    <TransactionsHistoryOnQuerySuccess data={data} title="Test Title" className="test-class" />,
+  )
   expect(container).toMatchSnapshot()
 })
